pyproject.toml
[project]
name = "lean-bindgen"
version = "0.1.0"
description = "Generate Lean 4 FFI bindings from C headers"
requires-python = ">=3.11"
dependencies = ["libclang>=16.0"]

[project.scripts]
lean-bindgen = "bindgen.__main__:main"

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bindgen"]
